Adding availity to create volumes
The create_volume action creates a volume of
given size in the specified availability zone.
A snapshot can be given to base the new
volume off of.

This commit also adds the ability to have a
volume created during the create action for
cloud provisioner and attach the volume to
the newly created instance.